to move sidebar component inside 'pages/layout/'
