+++ src/app/pages/layout/sidebar/sidebar.component.scss
@@ -10,10 +10,12 @@
     &.sidebar-mobile {
       background-color: rgba(255, 255, 255, 0.7);
       backdrop-filter: blur(16px);
+      border-right: none;
     }
 
     .sidebar-nav {
       padding: 15px;
+      margin-top: 65px;
       overflow: hidden;
 
       .sidebar-item {
@@ -78,8 +80,4 @@
       }
     }
   }
-}
-
-::ng-deep .mat-mdc-list-item.mat-mdc-list-item-interactive:hover::before {
-  background-color: transparent !important;
 }
