+++ src/app/utils/types/sidebar-component.type.ts
@@ -1,0 +1,3 @@
+export type UserType = 'user' | 'admin';
+
+export type SidenavMode = 'side' | 'over';
